bot 2.py
import os
from flask import Flask, request
from telegram import Update, Bot
from telegram.ext import Dispatcher, MessageHandler, filters
import requests
from bs4 import BeautifulSoup

TOKEN = os.environ['BOTTOKEN']
bot = Bot(token=TOKEN)
app = Flask(__name__)
dispatcher = Dispatcher(bot, None, workers=0, use_context=True)
WEBHOOK_URL = os.environ['WEBHOOK_URL']

# === Funções de scraping ===
# (Cole aqui as funções completas de Amazon, Shopee, ML e Magalu)

def resumo_oferta(url):
    # (Função que chama o scraping correto e monta a resposta)
    return "Exemplo de resposta formatada"

def responder(update: Update, _: object):
    url = update.message.text.strip()
    msg = resumo_oferta(url)
    update.message.reply_text(msg)

dispatcher.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, responder))

@app.route(f'/{TOKEN}', methods=['POST'])
def webhook():
    dispatcher.process_update(Update.de_json(request.get_json(force=True), bot))
    return 'OK'

@app.route('/')
def root():
    return 'Bot rodando!'

if __name__ == '__main__':
    bot.set_webhook(WEBHOOK_URL + TOKEN)
    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 8080)))
